+++ CHANGELOG.md
@@ -2,6 +2,32 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
+## 1.0.0alpha1 - 2018-02-26
+
+### Added
+
+- [#4](https://github.com/zendframework/zend-expressive-authentication-zendauthentication/pull/4)
+  adds support for the 1.0.0alpha3 (and later) release of zend-expressive-authentication.
+
+### Changed
+
+- Nothing.
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- [#4](https://github.com/zendframework/zend-expressive-authentication-zendauthentication/pull/4)
+  drops support for pre-1.0.0alpha3 releases of zend-expressive-authentication.
+
+- Nothing.
+
+### Fixed
+
+- Nothing.
+
 ## 0.3.0 - 2018-02-26
 
 ### Added